重构到一半，存个档

+++ cvAT_dllTest/cvAT_dllTest.cpp
@@ -21,9 +21,9 @@ int main()
 	// 初始化
 	//init();
 	// 准备变量
-	float x = 0;
-	float y = 0;
-	float a = 0;
+	double x = 0;
+	double y = 0;
+	double a = 0;
 	double x2 = 0;
 	double y2 = 0;
 	double a2 = 0;
